fix(desktop): extract chooseWindowsHermesHome as typed helper — review feedback
Per teknium1's review on #40233: extract the decision table into a typed
helper (hermes-home.ts) and call it from resolveHermesHome().

Also add the missing test case austinpickett noted:
- LOCALAPPDATA dir missing entirely but legacy has state.db → legacy wins

9/9 tests passing.
